refactor: split config into multiple modules
