+++ crates/utils/src/config/mod.rs
@@ -204,6 +204,7 @@ mod tests {
             Duration::from_secs(50),
             default_retry_count(),
             default_fixed_backoff(),
+            default_client_id_keep_alive_interval(),
         );
 
         let server_timeout = ServerTimeout::new(
@@ -305,7 +306,7 @@ mod tests {
     #[test]
     fn test_xline_server_default_config_should_be_loaded() {
         let config: XlineServerConfig = toml::from_str(
-            r#"[cluster]
+            "[cluster]
                 name = 'node1'
                 is_leader = true
                 peer_listen_urls = ['127.0.0.1:2380']
@@ -337,7 +338,7 @@ mod tests {
                 [auth]
 
                 [tls]
-                "#,
+                ",
         )
         .unwrap();
 
@@ -394,7 +395,7 @@ mod tests {
     #[test]
     fn test_auto_revision_compactor_config_should_be_loaded() {
         let config: XlineServerConfig = toml::from_str(
-            r#"[cluster]
+            "[cluster]
                 name = 'node1'
                 is_leader = true
                 peer_listen_urls = ['127.0.0.1:2380']
@@ -430,7 +431,7 @@ mod tests {
                 [auth]
 
                 [tls]
-                "#,
+                ",
         )
         .unwrap();
