+++ crates/utils/src/config/tls_config.rs
@@ -8,16 +8,16 @@ use serde::Deserialize;
 #[non_exhaustive]
 #[derive(Clone, Debug, Deserialize, PartialEq, Eq, Getters, Default)]
 pub struct TlsConfig {
-    /// The CA certificate file used by server to verify client certificates
+    /// The CA certificate file used by peer to verify client certificates
     #[getset(get = "pub")]
-    pub server_ca_cert_path: Option<PathBuf>,
-    /// The public key file used by server
+    pub peer_ca_cert_path: Option<PathBuf>,
+    /// The public key file used by peer
     #[getset(get = "pub")]
-    pub server_cert_path: Option<PathBuf>,
-    /// The private key file used by server
+    pub peer_cert_path: Option<PathBuf>,
+    /// The private key file used by peer
     #[getset(get = "pub")]
-    pub server_key_path: Option<PathBuf>,
-    /// The CA certificate file used by client to verify server certificates
+    pub peer_key_path: Option<PathBuf>,
+    /// The CA certificate file used by client to verify peer certificates
     #[getset(get = "pub")]
     pub client_ca_cert_path: Option<PathBuf>,
     /// The public key file used by client
@@ -33,20 +33,27 @@ impl TlsConfig {
     #[must_use]
     #[inline]
     pub fn new(
-        server_ca_cert_path: Option<PathBuf>,
-        server_cert_path: Option<PathBuf>,
-        server_key_path: Option<PathBuf>,
+        peer_ca_cert_path: Option<PathBuf>,
+        peer_cert_path: Option<PathBuf>,
+        peer_key_path: Option<PathBuf>,
         client_ca_cert_path: Option<PathBuf>,
         client_cert_path: Option<PathBuf>,
         client_key_path: Option<PathBuf>,
     ) -> Self {
         Self {
-            server_ca_cert_path,
-            server_cert_path,
-            server_key_path,
+            peer_ca_cert_path,
+            peer_cert_path,
+            peer_key_path,
             client_ca_cert_path,
             client_cert_path,
             client_key_path,
         }
+    }
+    
+    /// Whether the server tls is enabled
+    #[must_use]
+    #[inline]
+    pub fn server_tls_enabled(&self) -> bool {
+        self.peer_cert_path.is_some() && self.peer_key_path.is_some()
     }
 }
